🎯🗃️ 로그인된 유저와 작성자가 일치하는 글만 삭제 하게 구현 - `findOne`으로 찾는 데이터가 맞는지 확인후 삭제 or 에러

+++ src/pages/api/post/delete.js
@@ -1,11 +1,22 @@
 import { connectDB } from "@/utils/database.js";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth"
+import { authOptions } from "../auth/[...nextauth].js";
 
 export default async function handler(req, res) {
   if (req.method === 'DELETE') {
+    const session = await getServerSession(req, res, authOptions);
+
     const db = (await connectDB).db('forum');
-    let result = await db.collection('post').deleteOne({ _id: new ObjectId(req.body) })
+    const targetPost = await db.collection('post').findOne({ _id: new ObjectId(req.body) })
+    // console.log(targetPost)
+
+    if (targetPost.author == session.user.email) {
+      let result = await db.collection('post').deleteOne({ _id: new ObjectId(req.body) });
     // console.log(result) // 삭제결과 출력
-    res.status(200).json('삭제완료')
+      return res.status(200).json('삭제완료')
+    } else {
+      return res.status(500).json('현재유저가 작성한 글이 아닙니다.')
+    }
   }
 };
